rename storeapi to watchapi
